Make Oauth2 service compatible with DAPS and create integration test for token retrieval/validation from a test DAPS instance
