+++ extensions/iam/oauth2/oauth2-core/src/main/java/org/eclipse/dataspaceconnector/iam/oauth2/core/impl/JwkKey.java
@@ -1,0 +1,95 @@
+/*
+ *  Copyright (c) 2020, 2021 Microsoft Corporation
+ *
+ *  This program and the accompanying materials are made available under the
+ *  terms of the Apache License, Version 2.0 which is available at
+ *  https://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  SPDX-License-Identifier: Apache-2.0
+ *
+ *  Contributors:
+ *       Microsoft Corporation - initial API and implementation
+ *
+ */
+
+package org.eclipse.dataspaceconnector.iam.oauth2.core.impl;
+
+import java.util.List;
+
+/**
+ * Models an identity provider JWK key.
+ */
+public class JwkKey {
+    private String kty;
+    private String use;
+    private String kid;
+    private String x5t;
+    private String nn;
+    private String ee;
+    private List<String> x5c;
+    private String issuer;
+
+    public String getKty() {
+        return kty;
+    }
+
+    public void setKty(String kty) {
+        this.kty = kty;
+    }
+
+    public String getUse() {
+        return use;
+    }
+
+    public void setUse(String use) {
+        this.use = use;
+    }
+
+    public String getKid() {
+        return kid;
+    }
+
+    public void setKid(String kid) {
+        this.kid = kid;
+    }
+
+    public String getX5t() {
+        return x5t;
+    }
+
+    public void setX5t(String x5t) {
+        this.x5t = x5t;
+    }
+
+    public String getN() {
+        return nn;
+    }
+
+    public void setN(String n) {
+        nn = n;
+    }
+
+    public String getE() {
+        return ee;
+    }
+
+    public void setE(String e) {
+        ee = e;
+    }
+
+    public List<String> getX5c() {
+        return x5c;
+    }
+
+    public void setX5c(List<String> x5c) {
+        this.x5c = x5c;
+    }
+
+    public String getIssuer() {
+        return issuer;
+    }
+
+    public void setIssuer(String issuer) {
+        this.issuer = issuer;
+    }
+}
