+++ extensions/iam/oauth2/oauth2-spi/src/main/java/org/eclipse/dataspaceconnector/iam/oauth2/spi/JwtDecorator.java
@@ -1,0 +1,22 @@
+/*
+ *  Copyright (c) 2020, 2021 Microsoft Corporation
+ *
+ *  This program and the accompanying materials are made available under the
+ *  terms of the Apache License, Version 2.0 which is available at
+ *  https://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  SPDX-License-Identifier: Apache-2.0
+ *
+ *  Contributors:
+ *       Amadeus - initial API and implementation
+ *
+ */
+
+package org.eclipse.dataspaceconnector.iam.oauth2.spi;
+
+import com.nimbusds.jose.JWSHeader;
+import com.nimbusds.jwt.JWTClaimsSet;
+
+public interface JwtDecorator {
+    void decorate(JWSHeader.Builder header, JWTClaimsSet.Builder claimsSet);
+}
