+++ extensions/iam/oauth2/oauth2-spi/src/main/java/org/eclipse/dataspaceconnector/iam/oauth2/spi/ValidationRule.java
@@ -1,0 +1,29 @@
+/*
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.eclipse.dataspaceconnector.iam.oauth2.spi;
+
+import com.nimbusds.jwt.JWTClaimsSet;
+
+@FunctionalInterface
+public interface ValidationRule {
+
+    /**
+     * Rule that a token must satisfy in order to be valid.
+     *
+     * @param toVerify The Claims to verify.
+     * @param audience Required audience to validate the rule.
+     * @return ValidationRuleResult
+     */
+    ValidationRuleResult checkRule(JWTClaimsSet toVerify, String audience);
+}
