--- extensions/iam/oauth2/src/main/java/org/eclipse/dataspaceconnector/iam/oauth2/impl/JwkKeys.java
@@ -1,35 +1,0 @@
-/*
- *  Copyright (c) 2020, 2021 Microsoft Corporation
- *
- *  This program and the accompanying materials are made available under the
- *  terms of the Apache License, Version 2.0 which is available at
- *  https://www.apache.org/licenses/LICENSE-2.0
- *
- *  SPDX-License-Identifier: Apache-2.0
- *
- *  Contributors:
- *       Microsoft Corporation - initial API and implementation
- *
- */
-
-package org.eclipse.dataspaceconnector.iam.oauth2.impl;
-
-import org.jetbrains.annotations.Nullable;
-
-import java.util.List;
-
-/**
- * Models an identity provider JWK keys response.
- */
-public class JwkKeys {
-    private List<JwkKey> keys;
-
-    @Nullable
-    public List<JwkKey> getKeys() {
-        return keys;
-    }
-
-    public void setKeys(List<JwkKey> keys) {
-        this.keys = keys;
-    }
-}
