--- extensions/iam/oauth2/src/main/resources/META-INF/services/org.eclipse.dataspaceconnector.spi.system.ServiceExtension
@@ -1,15 +1,0 @@
-#
-#  Copyright (c) 2020, 2021 Microsoft Corporation
-#
-#  This program and the accompanying materials are made available under the
-#  terms of the Apache License, Version 2.0 which is available at
-#  https://www.apache.org/licenses/LICENSE-2.0
-#
-#  SPDX-License-Identifier: Apache-2.0
-#
-#  Contributors:
-#       Microsoft Corporation - initial API and implementation
-#
-#
-
-org.eclipse.dataspaceconnector.iam.oauth2.Oauth2Extension
