--- extensions/iam/oauth2/src/test/java/org/eclipse/dataspaceconnector/iam/oauth2/impl/FingerprintTest.java
@@ -1,64 +1,0 @@
-/*
- *  Copyright (c) 2020, 2021 Microsoft Corporation
- *
- *  This program and the accompanying materials are made available under the
- *  terms of the Apache License, Version 2.0 which is available at
- *  https://www.apache.org/licenses/LICENSE-2.0
- *
- *  SPDX-License-Identifier: Apache-2.0
- *
- *  Contributors:
- *       Microsoft Corporation - initial API and implementation
- *
- */
-
-package org.eclipse.dataspaceconnector.iam.oauth2.impl;
-
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
-
-import java.io.ByteArrayInputStream;
-import java.security.cert.CertificateException;
-import java.security.cert.CertificateFactory;
-import java.security.cert.X509Certificate;
-import java.util.Base64;
-
-import static org.junit.jupiter.api.Assertions.assertEquals;
-
-class FingerprintTest {
-    X509Certificate certificate;
-
-    @Test
-    void verifySha1HexFingerprint() throws Exception {
-        String fingerprint = Fingerprint.sha1HexFingerprint(certificate.getEncoded());
-        assertEquals("e1b2b8e5ec19d345a4f9afec8694f0d9c0aa25cf", fingerprint); // expected SHA1 fingerprint of the test cert
-    }
-
-    @Test
-    void verifySha1Base64Fingerprint() throws Exception {
-        String fingerprint = Fingerprint.sha1Base64Fingerprint(certificate.getEncoded());
-        assertEquals("4bK45ewZ00Wk+a/shpTw2cCqJc8=", fingerprint); // expected SHA1 fingerprint of the test cert
-    }
-
-    @BeforeEach
-    void setUp() throws CertificateException {
-        CertificateFactory factory = CertificateFactory.getInstance("X.509");
-        certificate = (X509Certificate) factory.generateCertificate(new ByteArrayInputStream(Base64.getDecoder().decode(X509_CERTIFICATE.getBytes())));
-    }
-
-    // X.509 cert with header and footer stripped (required by CertificateFactory)
-    static String X509_CERTIFICATE = "MIICVjCCAb8CAg37MA0GCSqGSIb3DQEBBQUAMIGbMQswCQYDVQQGEwJKUDEOMAwG" +
-            "A1UECBMFVG9reW8xEDAOBgNVBAcTB0NodW8ta3UxETAPBgNVBAoTCEZyYW5rNERE" +
-            "MRgwFgYDVQQLEw9XZWJDZXJ0IFN1cHBvcnQxGDAWBgNVBAMTD0ZyYW5rNEREIFdl" +
-            "YiBDQTEjMCEGCSqGSIb3DQEJARYUc3VwcG9ydEBmcmFuazRkZC5jb20wHhcNMTIw" +
-            "ODIyMDUyNzIzWhcNMTcwODIxMDUyNzIzWjBKMQswCQYDVQQGEwJKUDEOMAwGA1UE" +
-            "CAwFVG9reW8xETAPBgNVBAoMCEZyYW5rNEREMRgwFgYDVQQDDA93d3cuZXhhbXBs" +
-            "ZS5jb20wgZ8wDQYJKoZIhvcNAQEBBQADgY0AMIGJAoGBAMYBBrx5PlP0WNI/ZdzD" +
-            "+6Pktmurn+F2kQYbtc7XQh8/LTBvCo+P6iZoLEmUA9e7EXLRxgU1CVqeAi7QcAn9" +
-            "MwBlc8ksFJHB0rtf9pmf8Oza9E0Bynlq/4/Kb1x+d+AyhL7oK9tQwB24uHOueHi1" +
-            "C/iVv8CSWKiYe6hzN1txYe8rAgMBAAEwDQYJKoZIhvcNAQEFBQADgYEAASPdjigJ" +
-            "kXCqKWpnZ/Oc75EUcMi6HztaW8abUMlYXPIgkV2F7YanHOB7K4f7OOLjiz8DTPFf" +
-            "jC9UeuErhaA/zzWi8ewMTFZW/WshOrm3fNvcMrMLKtH534JKvcdMg6qIdjTFINIr" +
-            "evnAhf0cwULaebn+lMs8Pdl7y37+sfluVok=";
-
-}
