--- extensions/iam/oauth2/src/test/resources/readme-keys.txt
@@ -1,3 +1,0 @@
-Test cert alias is testcert1
-Test cert passwords are test123
-Test cert passphrase is test123
