+++ settings.gradle.kts
@@ -72,7 +72,9 @@ include(":extensions:in-memory:fcc-catalogquery-adapter-registry-memory")
 include(":extensions:in-memory:fcc-query-adapter-registry-memory")
 include(":extensions:in-memory:fcc-store-memory")
 include(":extensions:iam:iam-mock")
-include(":extensions:iam:oauth2")
+include(":extensions:iam:oauth2:oauth2-spi")
+include(":extensions:iam:oauth2:oauth2-core")
+include(":extensions:iam:daps")
 include(":extensions:iam:distributed-identity")
 include(":extensions:iam:distributed-identity:identity-did-spi")
 include(":extensions:iam:distributed-identity:identity-did-core")
